third_party/blink/web_tests/virtual/fetch-later/external/wpt/fetch/fetch-later/basic.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS fetchLater() cannot be called without request.
PASS fetchLater()'s return tells the deferred request is not yet sent.
Harness: the test ran to completion.
